[NFC][SYCL] Add `sycl/detail/fwd/accessor.hpp` helper
